fast_api_yatube/requirements.txt
asgiref==3.3.4
Django==3.2.3
django-ninja==0.13.0
flake8==3.9.2
mccabe==0.6.1
Pillow==8.2.0
pycodestyle==2.7.0
pydantic==1.8.2
pyflakes==2.3.1
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
